moved from services package in book details module
